requirements.txt
azure-cosmosdb-table
